update join the team page
