--- src/components/Pages/Pages.module.scss
@@ -1,53 +1,0 @@
-$fullViewportHeight: 100vh; // Represents the full height of the viewport
-$header-height: 76px; // Height of the header of the page
-$verticalspacing: 32px; // vertical space buffer
-$footer-height: 50px; // Height of the footer
-$vertical-spacing-buffer: 25px; // Provides spacing between the thank you contact form and the nav bar
-$extra-space: 25px; // Extra vertical space for the contact form
-
-.page {
-  overflow-y: auto;
-  padding-right: 10px;
-  flex: 1;
-  max-height: 55vh;
-}
-
-.pageHeader {
-  margin: 0;
-  margin-bottom: 40px;
-  color: #60718c;
-  font-family: 'Inter', sans-serif;
-  font-weight: 600;
-  font-size: 24px;
-  line-height: 30px;
-}
-
-.pageSubheader {
-  margin: 0;
-  margin-bottom: 10px;
-  color: #2d3748;
-  font-family: 'Inter', sans-serif;
-  font-weight: 600;
-  font-size: 20px;
-  line-height: 24px;
-}
-
-.pageText {
-  margin: 0;
-  margin-bottom: 20px;
-  color: #2d3748;
-  font-family: 'Inter', sans-serif;
-  font-weight: 500;
-  font-size: 14px;
-  line-height: 24px;
-
-  &Label {
-    margin-bottom: 0;
-    font-size: 16px;
-  }
-}
-
-.pageLink {
-  color: #033e8c;
-  text-decoration: none;
-}
